requirements.txt
python-dotenv>=1.0.0
pytz>=2024.1
requests>=2.31.0
python-dateutil>=2.8.2
yfinance>=0.2.0

